- fixed some glitches of the level edges. - Player can enter a level, but only the figure of the map is seen on it - WIP are the level physics
git-svn-id: https://clonekeenplus.svn.sourceforge.net/svnroot/clonekeenplus/cgenius/trunk@468 4df4b0f3-56ce-47cb-b001-ed939b7d65a6

+++ src/common/CMap.h
@@ -18,7 +18,7 @@ public:
 	CMap( SDL_Surface *p_scrollsurface, CTilemap *p_Tilemap);
 	virtual ~CMap();
 
-	bool gotoPos( Uint16 x, Uint16 y );
+	bool gotoPos( int x, int y );
 	void scrollLeft(void);
 	void scrollRight(void);
 	void scrollUp(void);
